move interpolation function from align to core
